requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.2.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
openai>=0.27.0
anthropic>=0.3.0
janome>=0.4.2
sudachipy>=0.6.7
sudachidict-core>=20230110
tqdm>=4.65.0
pyyaml>=6.0
scipy>=1.10.0
statsmodels>=0.14.0
jupyter>=1.0.0
